+++ CHANGELOG.md
@@ -13,8 +13,11 @@ in this version I will refactoring the code & make it enhanced.
 **ADDED**
 
 -   add prettier standard code for react : it's will help keep code formated in correct way.
+-   add babel proposal class properties for make auto bind & auto class properties assignments.
 
 **CHANGED**
 
 -   correct variable declaration from `var` to `let` in all files, because React work in strict mode.
+-   change declaration of `let` to `const` for objects.
 -   change files `utils.js` some with enhanced and clear expression + change not efficiency loops by array itterators functions.
+-   change in class constructor that's not required.